+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 0.0.1
+# Changelog
 
-* TODO: Describe initial release.
+## [0.9.0] - 2020-07-04
+### Added
+- **Flutter:** Tracker initialization, Subject setter and event tracking
+- **Android:** Tracker, Emitter and Subject creation
+- **Android:** PageView, Structured, Unstructured, ScreenView, ConsentWithdrawn, ConsentGranted, Timing and EcommTransaction event tracking
+- **iOS:** PageView, Structured, Unstructured, ScreenView, ConsentWithdrawn, ConsentGranted, Timing, EcommTransaction and PushNotification event tracking